Move most of the Windows code into a separate header
