+++ src/library/cmd_reader.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "impl/base.h"
+#include "impl/cmd_reader_windows.h"
 
 namespace xlang::cmd
 {
@@ -173,7 +174,7 @@ namespace xlang::cmd
 
                 if (path == "sdk" || path == "sdk+")
                 {
-                    sdk_version = get_sdk_version();
+                    sdk_version = impl::get_sdk_version();
                 }
                 else
                 {
@@ -188,13 +189,13 @@ namespace xlang::cmd
 
                 if (!sdk_version.empty())
                 {
-                    auto sdk_path = get_sdk_path();
+                    auto sdk_path = impl::get_sdk_path();
                     auto xml_path = sdk_path;
                     xml_path /= L"Platforms\\UAP";
                     xml_path /= sdk_version;
                     xml_path /= L"Platform.xml";
 
-                    add_files_from_xml(files, sdk_version, xml_path, sdk_path);
+                    impl::add_files_from_xml(files, sdk_version, xml_path, sdk_path);
 
                     if (path.back() != '+')
                     {
@@ -206,7 +207,7 @@ namespace xlang::cmd
                         xml_path = item.path() / sdk_version;
                         xml_path /= L"SDKManifest.xml";
 
-                        add_files_from_xml(files, sdk_version, xml_path, sdk_path);
+                        impl::add_files_from_xml(files, sdk_version, xml_path, sdk_path);
                     }
 
                     continue;
@@ -224,246 +225,6 @@ namespace xlang::cmd
         }
 
     private:
-
-#if XLANG_PLATFORM_WINDOWS
-
-        struct registry_key
-        {
-            HKEY handle{};
-
-            registry_key(registry_key const&) = delete;
-            registry_key& operator=(registry_key const&) = delete;
-
-            ~registry_key() noexcept
-            {
-                if (handle)
-                {
-                    RegCloseKey(handle);
-                }
-            }
-        };
-
-        template <typename T>
-        struct com_ptr
-        {
-            T* ptr{};
-
-            com_ptr(com_ptr const&) = delete;
-            com_ptr& operator=(com_ptr const&) = delete;
-
-            com_ptr() noexcept = default;
-
-            ~com_ptr() noexcept
-            {
-                if (ptr)
-                {
-                    ptr->Release();
-                }
-            }
-
-            auto operator->() const noexcept
-            {
-                return ptr;
-            }
-        };
-
-        static void check_xml(HRESULT result)
-        {
-            if (result < 0)
-            {
-                throw_invalid("Could not read the Windows SDK's Platform.xml");
-            }
-        }
-
-        static void add_files_from_xml(
-            std::set<std::string>& files,
-            std::string const& sdk_version,
-            std::experimental::filesystem::path const& xml_path,
-            std::experimental::filesystem::path const& sdk_path)
-        {
-            com_ptr<IStream> stream;
-
-            check_xml(SHCreateStreamOnFileW(
-                xml_path.c_str(),
-                STGM_READ, &stream.ptr));
-
-            com_ptr<IXmlReader> reader;
-
-            check_xml(CreateXmlReader(
-                __uuidof(IXmlReader),
-                reinterpret_cast<void**>(&reader.ptr),
-                nullptr));
-
-            check_xml(reader->SetInput(stream.ptr));
-            XmlNodeType node_type = XmlNodeType_None;
-
-            while (S_OK == reader->Read(&node_type))
-            {
-                if (node_type != XmlNodeType_Element)
-                {
-                    continue;
-                }
-
-                wchar_t const* value{ nullptr };
-                check_xml(reader->GetLocalName(&value, nullptr));
-
-                if (0 != wcscmp(value, L"ApiContract"))
-                {
-                    continue;
-                }
-
-                auto path = sdk_path;
-                path /= L"References";
-                path /= sdk_version;
-
-                check_xml(reader->MoveToAttributeByName(L"name", nullptr));
-                check_xml(reader->GetValue(&value, nullptr));
-                path /= value;
-
-                check_xml(reader->MoveToAttributeByName(L"version", nullptr));
-                check_xml(reader->GetValue(&value, nullptr));
-                path /= value;
-
-                check_xml(reader->MoveToAttributeByName(L"name", nullptr));
-                check_xml(reader->GetValue(&value, nullptr));
-                path /= value;
-
-                path += L".winmd";
-                files.insert(path.string());
-            }
-        }
-
-        static registry_key open_sdk()
-        {
-            HKEY key;
-
-            if (0 != RegOpenKeyEx(
-                HKEY_LOCAL_MACHINE,
-                "SOFTWARE\\Microsoft\\Windows Kits\\Installed Roots",
-                0,
-                KEY_READ,
-                &key))
-            {
-                throw_invalid("Could not find the Windows SDK");
-            }
-
-            return { key };
-        }
-
-        static std::experimental::filesystem::path get_sdk_path()
-        {
-            auto key = open_sdk();
-
-            DWORD path_size = 0;
-
-            if (0 != RegQueryValueExW(
-                key.handle,
-                L"KitsRoot10",
-                nullptr,
-                nullptr,
-                nullptr,
-                &path_size))
-            {
-                throw_invalid("Could not find the Windows SDK");
-            }
-
-            std::wstring root((path_size / sizeof(wchar_t)) - 1, L'?');
-
-            RegQueryValueExW(
-                key.handle,
-                L"KitsRoot10",
-                nullptr,
-                nullptr,
-                reinterpret_cast<BYTE*>(root.data()),
-                &path_size);
-
-            return root;
-        }
-
-        static std::string get_module_path()
-        {
-            std::string path(100, '?');
-            DWORD actual_size{};
-
-            while (true)
-            {
-                actual_size = GetModuleFileNameA(nullptr, path.data(), 1 + static_cast<uint32_t>(path.size()));
-
-                if (actual_size < 1 + path.size())
-                {
-                    path.resize(actual_size);
-                    break;
-                }
-                else
-                {
-                    path.resize(path.size() * 2, '?');
-                }
-            }
-
-            return path;
-        }
-
-        static std::string get_sdk_version()
-        {
-            auto module_path = get_module_path();
-            std::regex rx(R"(((\d+)\.(\d+)\.(\d+)\.(\d+)))");
-            std::cmatch match;
-
-            if (std::regex_search(module_path.c_str(), match, rx))
-            {
-                return match[1].str();
-            }
-
-            auto key = open_sdk();
-            uint32_t index{};
-            std::array<char, 100> subkey;
-            std::array<unsigned long, 4> version_parts{};
-            std::string result;
-
-            while (0 == RegEnumKeyA(key.handle, index++, subkey.data(), static_cast<uint32_t>(subkey.size())))
-            {
-                if (!std::regex_match(subkey.data(), match, rx))
-                {
-                    continue;
-                }
-
-                char* next_part = subkey.data();
-
-                for (size_t i = 0; ; ++i)
-                {
-                    auto version_part = strtoul(next_part, &next_part, 10);
-
-                    if (version_part < version_parts[i])
-                    {
-                        break;
-                    }
-
-                    version_parts[i] = version_part;
-
-                    if (i == std::size(version_parts) - 1)
-                    {
-                        result = subkey.data();
-                        break;
-                    }
-
-                    if (!next_part)
-                    {
-                        break;
-                    }
-
-                    ++next_part;
-                }
-            }
-
-            if (result.empty())
-            {
-                throw_invalid("Could not find the Windows SDK");
-            }
-
-            return result;
-        }
-
-#endif
 
         std::vector<option>::const_iterator find(std::vector<option> const& options, std::string_view const& arg)
         {
